refactor (input field): use field base and move to form path
