--- lib/components/InputField/styles.scss
@@ -1,134 +1,0 @@
-.au-input {
-  cursor: pointer;
-
-  &__field {
-    border: 1px solid $color-neutral-50;
-    background-color: $color-neutral-00;
-    border-radius: 8px;
-    font-size: 16px;
-    line-height: 22px;
-    padding: 16px;
-    width: 100%;
-
-    &:hover {
-      border: 1px solid $color-neutral-70;
-    }
-
-    &:focus {
-      border: 2px solid $color-brand-blue-40;
-    }
-  }
-
-  &:focus-within:not(.au-input--error) {
-    [class*='__header-label'] {
-      color: $color-brand-blue-40;
-    }
-  }
-
-  &__header {
-    display: flex;
-    justify-content: space-between;
-    padding-bottom: 8px;
-
-    &-label {
-      color: $color-neutral-50;
-      cursor: pointer;
-      font-size: 14px;
-      font-weight: 600;
-      line-height: 22px;
-
-      &--required {
-        color: $color-error-50;
-      }
-    }
-
-    &-icon {
-      span {
-        font-size: 14px;
-        color: $color-neutral-50;
-      }
-    }
-  }
-
-  &__error-message {
-    color: $color-error-50;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 24px;
-    padding-top: 8px;
-  }
-
-  &--disabled {
-    cursor: not-allowed;
-
-    [class*='__field'] {
-      background-color: $color-neutral-20;
-      border-color: $color-neutral-40;
-      color: $color-neutral-40;
-      cursor: not-allowed;
-    }
-
-    [class*='__header-label'] {
-      color: $color-neutral-40;
-      cursor: not-allowed;
-    }
-  }
-
-  &--error {
-    [class*='__field'] {
-      background-color: $color-error-00;
-      border-color: $color-error-50;
-    }
-
-    [class*='__header-label'] {
-      color: $color-error-50;
-    }
-  }
-
-  &--success {
-    [class*='__field'] {
-      animation: inputSuccess 2s forwards;
-    }
-
-    [class*='__header-label'] {
-      animation: labelSuccess 2s forwards;
-    }
-
-    [class*='__header-icon'] {
-      .au-icon {
-        animation: iconSucess 2s forwards;
-      }
-    }
-  }
-
-  @keyframes inputSuccess {
-    0%,
-    70% {
-      background-color: $color-success-00;
-      border-color: $color-success-50;
-    }
-    100% {
-      @extend [class*='__field'];
-    }
-  }
-
-  @keyframes labelSuccess {
-    0%,
-    70% {
-      color: $color-success-50;
-    }
-    100% {
-      @extend [class*='__header-label'];
-    }
-  }
-
-  @keyframes iconSucess {
-    0%,
-    70% {
-      opacity: 1;
-    }
-    100% {
-      opacity: 0;
-    }
-  }
-}
