+++ lib/components/form/InputField/InputField.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { InputField } from './index'
 
 const meta: Meta<typeof InputField> = {
-  title: 'Components/InputField',
+  title: 'Components/form/InputField',
   component: InputField,
   tags: ['autodocs'],
 }
