+++ lib/main.ts
@@ -12,7 +12,7 @@ export { NotificationsBar } from './components/NotificationsBar'
 export { ProfileNav } from './components/ProfileNav'
 export { Text } from './components/Text'
 export { Footer } from './components/Footer'
-export { InputField } from './components/InputField'
+export { InputField } from './components/form/InputField'
 
 // Hooks
 export { useDrawer } from './components/Drawer/hooks'
